rrr: convert rpc/request_queue to module (rrr.request_queue)
request_queue.cpp module; .hpp shim. Exports OverflowStrategy enum,
RequestQueueConfig (with 4 presets), QueuedRequest POD, RequestQueue
class (SpinMutex<rusty::VecDeque<QueuedRequest>>).

Imports rrr.basetypes, rrr.marshal, rrr.threading.

32 modules. 67.97 s wallclock (4.03× baseline).
