+++ src/rrr/CMakeLists.txt
@@ -47,6 +47,7 @@ set(RRR_MODULE_SRC
     ${CMAKE_CURRENT_SOURCE_DIR}/rpc/pollable_proxy.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/rpc/reconnect_policy.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/rpc/request_options.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/rpc/request_queue.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/rpc/utils.cpp
 )
 list(REMOVE_ITEM RRR_BENCH_SRC ${RRR_MODULE_SRC})
